use doubles and fk to exercise name

+++ gt-core/src/entities_backend/user_info_ts.rs
@@ -9,10 +9,10 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub user_id: i32,
-    pub height: Option<f32>,
-    pub weight: Option<f32>,
-    pub muscle_mass: Option<f32>,
-    pub body_fat: Option<f32>,
+    pub height: Option<f64>,
+    pub weight: Option<f64>,
+    pub muscle_mass: Option<f64>,
+    pub body_fat: Option<f64>,
     pub created_at: DateTime,
 }
